Homework3 fixed
